Courses/Agents/CrewAi/crew-agents/pyproject.toml
[tool.poetry]
name = "simple-agent"
version = "0.1.0"
description = "A simple CrewAI agent that answers questions"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.83.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
simple-agent = "simple_agent.main:run"